Wrong error message in CSV importer (#51966)
* Test that sku skipping works

#46505

* Switch SKU and ID error detection

If we find an product with the same SKU we overwrite the imported product id with the existing one. When we check for the existing ID first, we would get now a match although the imported ID did not match. Therefore we first check if the SKU is already present before checking for the ID

#46505

* Add doc

* Add changelog

* Add changefile(s) from automation for the following project(s): woocommerce

* Remove unneeded changelog entry pulled in during rebase.

* Remove mistakenly pushed debug code.

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Barry Hughes <3594411+barryhughes@users.noreply.github.com>